New compara DB schema for multiple genomic alignments.
Mainly the old genomic_align_block table is splited into two tables, one for each alignment block (new genomic_align_block table) and another one for storing every sequence of every alignment block (new genomic_align table).  Since there is no more consensus and query genomes in the structure, the genomic_align_genome table is not in the new schema anymore. There is a new table called genomic_align_group which allows to group aligned sequences in several ways. The species_set concept is replaced by the method_link_species_set which relates every method_link with the species_set for which it has been used.

There are also a couple of minor changes in the dnafrag table. The start and end columns are replaced by the length of the dnafrag and the dnafrag_type is renamed as coord_system_name to follow EnsEMBL core naming system.