complete redesign of underlying tree/graph datastructure. The old nestedset data structure assumed that the data would be in the form of trees but this has proved not to be the case.  Most of the data is in the primary form of acyclic graphs (also called unrooted trees). Trees are directional, rooted, acyclic graphs. A new datastructure to represent any graph structure in object form has been developed.  Some functions have been added to give additional function when the graphs are acyclic.  NestedSet now inherits from Graph::Node The previous NestedSet API has been left in place, but the implementation now is done by calls to the underlying Graph API.
The central element in this datastructure is the Graph::Link
Each link is a global entity(not retained), but each link retains each
Node on either side of it.  The Link is inherently bi-directional so
graph traversal algorithms are easier to implement.  Each node keeps a list
of the Links internally that are attached to it.  Retain/Release is more
complex, but based around the concept of releasing links to release the
graph (function cascade_unlink for example). A cascade release is done
as a chain reaction of breaking links via a directional graph traversal.

The previous external API and function of NestedSet has been completely
retained but internal variables and hidden methods are changed.