Hi Bert,
I'm sending this to Maurice because he is implementing the GO display,
and to Ian, because you forwarded this ticket to him just now.

Maurice, code-wise this should be a simple fix where you change
"goslim_goa" to "goslim_generic".  On the web pages themselves, this
subset should be referred to as "The GOSlim Generic subset of GO".

Unless I've overlooked something, that should be all.

There's nothing for Ian to do with regards to this, as far as I can see.


Cheers,
Andreas