This time to head.
Update vega protein tree pipleine config to optimise to streamline vega
builds.

Reduces CPU time by 275hrs (47%) and clock time by 16h (64%).

1. time limit and automate ramping of memory resources for mcoffee jobs so
that it matches our old manual baby-sitting process, to mean pipe runs
automatically;

2. reduce parallelism in database-intensive tasks so that we don't kill our
poor database server; so that we don't waste farm cpu-time waiting for the
server; so that total clock-time is reduced due to less lock-furtling in
database server; and so we don't get failures from connection timeouts;

3. apply sensible cuttofs for tasks based on execution time to accurately
distinguish jobs which will never finish (determined by analysis of earlier
runs), so as not to waste CPU or wall time on endless;

4. more accurately matches requested resources to those used to speed up
allocation of slots on farm and to be nice to it when running and to deal
with memory limit changes.

Modifies the main protein pipeleine rather than duplicating it to
increase chance that modifications to pipeline by compara will "just work".

Slightly ugly is that without the ability to time-limit jobs, we have to
time-limit workers (via LSF runlimit) which means there's a small chance
that good jobs will fail leading to a suboptimal alignement. The probability
is very samll, though, less than one protein slightly sub-optimally aligned
in 10000 releases, way below the noise floor.