Prevent overlap in bed tracks due to all the rectangles everywhere in ensembl being one pixel too high/wide. Ideally we would fix that, but I'm not brave enough yet as it might break everything.
The underlying error comes from believing (0,0)-(10,10) is a 10x10 box
whereas actually it's 11x11, in gif.pm etc. This affects genes, contigs....
You don't usually notce because there's a gap, but not in bed features on
different strands which should abutt but had a one pixel overbite until
this fix.