Fix bug in SVG renderer by making it incorrect like all other renderers, sad face.
Most other renderers render rectangles one wider and one higher than
actually specified, SVG formerly rendered them the correct size. These
rectangles are created in hundreds of places in the drawing code so
subtracting one from them all isn't practical on a bugfix timescale,
but would make a good project for someone with time on their hands!

At some point, someone spotted this bug and worked around the issue by
creating the L shapes of condensed view genes of width 0. This works in
the broken renderers but is missing from SVG because it is (was)
implemented correctly.