Fix for trouble turning tracks on reported via bugs@ by EK on 2013-01-11.
The underlying issue was that we were using whether something was already in the tree or not to decide whether a subsequent add was a duplicate, in add_tracks in ImageConfig. Because we are loading some data early, the presence of a custom config already in the tree meant that even on the first call it wasn't added.

I've changed the deduplication to use a different mechanism, by just maintaining a hash of keys inside the ImageConfig object which should, I think, be the same thing: as long as create_track always adds to the tree, or the circumstances where it doesn't we don't care about; and we don't care about deduplication between calls to create_track via add_tracks and calls from outside add_tracks (eg DAS/userdata tracks).