Rewrite alignment_pairwise to use genomicalignblocks and fix lots of bugs (hopefully not introducting too many more), particularly with the display of CIGAR strings and the green "joining" regions.
A lot of extra code in here is because much of the drawing code (eg this
classes supertype) provide utility methods which we cannot use because
they assume we have featuresi, not genomicaligns. As we use genomic-aligns
more and more, we should push these methods up to a new genomicalign
superclass which will simplify this file and provide the methods for other
glyphsets.