Fixes to TrackHub attachment
Allow attachment of TrackHubs without a hub.txt - accept any URL that has a . or ? after the last slash.
URLs ending in / or /something_without_._or_? will be assumed to have a hub.txt (/hub.txt if not ending in /).
Set the name to be shortLabel from the hub file, unless otherwise specified.
Improved error handling and reporting when a hub fails to attach.