Fix checkboxes for compara multiway config.
add_element in FieldSet used to convert "selected" into "checked".
Now that's deprecated and we're stopping using it, we should allow
selected in Checkbox.pm as an alternative to checked (at least
until we can engineer all the code that generates params to emit
the correct key, if that's the way we want to go).