The (user-defined) species tree must be minimized
This comes from an EG bug: the species tree had a few nodes with a single
child. This causes problems later in the CAFESpeciesTree runnable, and
perhaps in other analysis / programs.

The commit adds a simple command to minimize the user-defined tree, and a
SQL healthcheck to ensure we don't go further in the pipeline