Tidying-up of content_sub_slice() arguments
content_sub_slice() does not deal with warnings any more (it was only
appending them to the HTML anyway). It is now up to the caller (in two
cases: content()) to create the HTML and add it to the page.
It was actually almost the case in Variation/Compara_Alignments.pm