Mice are not the only Mus
E.g. ferret (mustela putorius furo) and housefly (musca domestica)

Must be a better solution; CCDS flag in the meta table? Or maybe process the transcript data before building the table and only include the column if $ccds got populated - might not be a good solution if there is significance in showing absence of CCDS.