bugfix: isoforms in the same family
If the gene has several proteins in the same family, fetch_all was
returning an instance of the family object for each of them. With
the patch, only the unique list of families is returned.
A similar bug was affecting the number of families stored in the
gene_member table.