Reduced configration for reg summary page.
RT#414985

The bug was that wiggles don't display on this page despite being
configurable. It's probably not a good idea from a UI PoV to allow
wiggles to be turned on here as that confuses what pages are for.
Also it's a massive performance hit (~2s) to allow them to be turned on,
even if none are. So the decision was to not let them be configured on.