Removing click event from track description
The click event that displays the renderer popup was registered with the li.track element. Thus anything inside this li.track was displaying this popup when clicked, and this included the track description box.
To fix this behavior, a new element div.track_name containing just the track label is added inside li.track and click event and all the required CSS are moved to this new element.
Now, the popup is displayed only when clicked on the track name or the renderer icon.