Update FilterDuplicates.pm
Updated genomic_align_blocks_identical method logic (no more test on identical score).
Added a flag 'filter_duplicates_net' so we can additional filtering logic for the filter_duplicates_net step (in addition to filter identical blocks, also filter overlapping blocks). This flag is set to 0 by default, and needs to be overwritten accordingly in the PairAligner config file, if you want to set it on.