Use the API to fetch the toplevel slices.
At the first glance, one may say that the direct SQL query must be faster, but
as a matter of fact, because of the line below, each slice was then fetched
independently by its name which causes a major slowdown.

The simple solution is to use the core API to fetch all the toplevel slice objects.