Restructured MLSSIDFactory
The runnable now accepts a hashref 'methods' that associates each method_link
to a branch number. Only the methods registered there will be dataflown

It now creates jobs with the "homo_mlss_id" key instead of "mlss_id". This
allows other jobs to still access the pipeline-wide "mlss_id" of the gene-trees.