Fixing style for toggleable table's heading
Since it was float:left, when table is minimised, it moves the content following the table to the right of the heading.
The reason it was float:left is to accommodate the 'Back to top'  link within the same line - so the link is moved above the heading to be in-line with heading.