New structure for the Synteny pipeline
The pipeline is now using eHive's parameter-stack mechanism, i.e. the jobs
inherit the parameters of their parents.
The pipeline should be configured exclusively from the command line, with the
-pairwise_db_url and possibly -pairwise_mlss_id parameters. If the latter is
skipped, the pipeline will use all the pairwise alignments found on the server.
The pipeline automatically finds the alignments that are missing syntenies and
compute these (incl. the stats)
The analysis "compute_synteny_start" can be seeded multiple times