Added a sleep to avoid errors caused by delays on I/O.
Sometimes the file system can be a bit slow in writing the outputs to
disk. That could mislead the worker causing it to think that a
particular job has failed since it didn’t write any output. But that
glitch is only temporary, since the FS is already writing the file. By
adding a sleep we make sure to wait for that output to be ready.