Added: feature to deal with specified set of trees.
With this change we allow the user to define the reference clusterset_id
(ref_tree_clusterset), the alternative trees to be compared
(alternative_trees) and if the trees need to be re-rooted
(reroot_with_sdi).

The default mode of this module was kept as it was, the new features are
only in place when strictly specified by the user.