Separate config into separate module.
Lots of stuff broken: hyphenation and server-side sorting at least,
probably much more. DO NOT MERGE YET.

This work is to put all state dependent only on the config into a
module so that it can be used confidently as a key to a cache. This
fine-grained cache will be useful generally, but critical when it comes
to re-sorting on the serverside to avoid wrong-order returns during a
rapid filter-then-sort.