New runnable to store tree in the DB from a newick file.
Its useful to use when a particular tree needs to be inserted into the
database. e.g. when a very big tree was already computed and we dont
want to re-infer it, so it could be loaded from the newick straint into
the DB. It also uses the -output_clusterset_id as input.