Speed-up the pipeline by merging two analyses
Compare_orthologs was a very fast analysis but there can be millions of jobs,
which makes the pipeline terribly slow because of the job overhead. Also eHive
sometimes chokes on that amount of jobs.

Here we combine "create_comparison_job_arrays" and "check_ortholog_neighbors".
This is made easy by the fact that Comparison_job_arrays only had run().
Comparison_job_arrays is now a subclass of Compare_orthologs that calls
SUPER::run() instead of dataflowing.