New feature: random binarization, per node.
This piece of code takes advantage of the method random_binarize_node().
Allowing us to binarize only the multifurcated nodes, instead of passing
the whole tree to the method.
This is not just faster and more efficient but it is also more accurate
since it will not rearrange nodes that are already binary.