Bugfix: Pruning the right tree now.
Used when the taxonomic sub-tree has more species than the ones in
the gene-tree, in those cases we need to remove the extra leaves.
We before this commit we were pruning the $species_tree which isn't used
anywhere else from this point on.
The right tree to be pruned is the $castedMrca.