[ENSWEB-2259] Fixed a hack with another
Also fixes ENSWEB-884 and ENSWEB-2221.

The original hack in c240b93fa7278333812cbe577e500970bfb81ee6 was to display the tracks that are on by default in the trackhub by making the drawing code and the matrix code think that the user has turned those tracks on. It worked, but caused few problems.
1. Since it saved the data in the user config, the user was never able to turn those tracks off as it would overwrite those 'off's with 'on's.
2. The numbering on the track hubs matrix was wrong since it was only turning on the cell but not an individual track.

So the fix is to make the drawing code think the user turned on the individual tracks too, along with the cells - this fixes the numbering, but don't overwrite the user configs - this enables the user to turn those tracks off, and then just before saving the data to the user records, remove it from the record - this makes sure the defaults are not saved in the user configs.