Reverting these changes.
The best solution for this issue is to revert these commits.
There is no need to add Bio::EnsEMBL::Hive::RunnableDB::SqlHealthcheck
to every single runnable, but only to the RunnableDB/GeneTrees/SqlHealthChecks.pm
Since we're calling a methods from SqlHealthChecks, we shouldn't have to
add additional libraries, it should all be managed by SqlHealthChecks
itself.