Tidied Page controller
No need to 'return' after redirects - hub->redirect deals with it.
Get rid of toggle_ code complexity - not in use i.e. we don't need to toggle elements on the page acc. to a param in url anymore - if this functionality is needed, it should be implemented in JS using url anchor. Does it with url param might interfere with url based caching mechanisms