bugfix: current tree has more members than reused tree.
Adds members that are in the current tree but not in the reused tree.
Fixes very specific cases in which some members may me present in the
current tree but not in the reused tree.
They occur when there were no change in the sequence level (hence the
MD5 comparisons didn't catch it) but there were changes in the clusters
(when HMMs get updated).