New method to remap the seq_member_ids
It remaps the re_used tree seq_member_ids with the new ones used in the
current_tree. This is necessary for the cases where a species have been
updated e.g. (gene-set || assembly), meaning that the seq_members are
re-inserted hence not sharing the same ids with the reused database.