Reverting change in the schema.
This is a production only table and should be declared in the
sql/pipeline-tables.sql file. So it is now being removed from the
sql/table.sql file.
Additionally there is no need for adding a patch, since it is a
production table.

This reverts the following commits:
d534e404994038dc3a82ccef6044aa619a312fa1 and
303d15cc000d8ca1876d99fd9c9be715855058f2