Can now process several EPO2X alignments
Previously it would only consider the biggest one (number of species).

As a consequence, we don't need to store the low-coverage-assembly
species-set any-more. In fact, the NCRecoverEPO Runnable still had to
connect to the EPO dbs to find out which species were high-coverage. This
is all done by the runnable, now