ENSWEB-2756
We need to release! This seems to fix the bug based on evidence from
tracing variable contents. But I don't understand the code. We need to
talkto Michael and Anne about this when it's less chaotic. This fix may
well introduce bugs: I've not found any and they probably won't be as
serious as the one it fixes.