Use defer method in js.
After much reading this seems to be the best we can do while supporting
both some of the nasty things we do in our scripts and our usual range
of browsers.

There's a possibility that IE<10 has issues with defer due to an obscure
bug, but a quick survey suggests we don't hit it (requires DOM-manipulation
very early in load cycle to trigger, which we seem not to do).

If there are issues they should be easy enough to work around.