Avoid loop when querying for ontology availability
With current implementation, in order to decide whether to enable ontology related menu items
on the left hand side of a page (e.g. gene summary), it needs to fetch from ontology database
for related ontologies.

The lookup of ontology adaptor takes long time for Bacteria because the underlying implementaion
requires an exhaustive iteration through the whole registry:
https://github.com/Ensembl/ensembl/blob/release/88/modules/Bio/EnsEMBL/Registry.pm#L705-L714

Furthermore, the current implementation repeatly fetches from the database for the same query:
https://github.com/Ensembl/ensembl-webcode/blob/release/88/modules/EnsEMBL/Web/Query/Availability/Gene.pm#L282

This change tries to fetch from the ontology database for only once.