Remove htsobj cache once data is fetched from bam/cram
For some cram files, the htslib consumes large amount of memory to parse and fetch data
from them. This creates a problem for a location view that has multiple those cram tracks
configured. The memory will quickly exceed the process limit.

This change removes the htsobj cache once the data is fetched for a certain track which
effectively clean up the memory so that all the following tracks could have enough memory
to be processed.

Related to ENSEMBL-5035.