New method of detecting the inclusion of a subspecies and its species
The previous method was detecting these cases by comparing the size of the
tree. For it to work, it had to use release_children() which was causing
some harm (not sure exactly where).
The new method uses left_/right_index to detect that a node is below
another node. It doesn't need to fetch extra nodes / release nodes, so
should be safer to use.