bugfix: the previous mlss_id is not given any more. Need to query the homology_id_mapping table
Since all the dbIDs may be different (genome_db_id, gene_member_id), use
gene_member.stable_id to map members. This means that the whole mapping
(both genome_dbs) has to be loaded for each genome as a reference.

This might slow down the pipeline in its old structure, but as the goal is
to move to a single Runnable that processes both species, this is in fact
fine.