Allow the runnable to auto restart RAxML after 12 hours.
The idea here is to first try rerunning RAxML before moving it to a
higher capacity analysis. We have observed that in many cases RAxML
would be running for 4 days, and if we restart these jobs, they would
finish in less than 1 hour. If after restarting, RAxML still don't
finish in 12 hours we dataflow it to the next level capacity.