Switch to using opt for hdf5
Rather than hard-coding the path for hdf5 1.8 keg we can use Linuxbrew's 
opt directory to always use the latest build. Since this will always be the 1.8 release
there should be little to no reason why sub-release bumps will affect the HAL bindings. 
If they did we would have to rebuild the hal bindings in XS and in Linuxbrew.