New health check to account for orphan genes.
This HC allows us to define different ratio of orphan genes for
different taxon levels.
Since the pipeline computes these ratios at an early stage, the idea is
to make it fail if the value is too low for a certain species, and then
we may just decide to pull this genome out of the release.