bugfix: Add the component name to the display name in the object / table rather than appending it on the fly
Done with @mateuspatricio

The issue was that display_name() not returning what is stored in the
database, GenomeDBs could not be copied with fetch+store (which happens
when populating the new database). In the master database, all the wheat
genome_dbs were named Triticum Aestivum, whilst they had "component X"
appended in the populated release database.

Now, the component name is really recorded in the display name