bufix: this code was in the "else" but also applies to the "then" -> moved after the "if"
This caused an error in some cases in the Alignment Text view: when the
preview region (the first 200bp) is in a gap in the ancestral sequences,
the first "underlying slice" didn't have "_tree" and "_node_in_tree", so a
Perl error was raised in the webcode ("cannot run ... on an undefined
value").
Since both fields come from $this_slice, which is available in the "then"
part of the if, just moved the code to always set those fields if possible.