Defined the species-tree in ENV too
Note that for the MSA and species-tree pipelines I have removed the
internal option and the pipeline now directly use "binary_species_tree"
because it was used in just one place.
On the other hand, in the gene-tree pipelines, because the option was used
in many places, I have just changed the value of the existing option. This
means that there are now two options for the binary species-tree. I hope it
won't be too confusing.