Disabling disconnect_when_inactive.
Disconnecting was generating generating the error code 99, in which too
many connections are being made to a server.
This is a problem for species with lots of dnafrags, in which each one
of them will result in opening a connection.
A ticket to redesing this part of the pipeline was opened, in the
meantime we will disable the disconnect_when_inactive.