Better die message
I am investigating further as a failing job instructed me. Knowing the proportion of my orphaned genes, in addition to knowing the pipeline thinks it's too many, helped me judge that the failure is due to my crappy data, as opposed to the pipeline run going wrong - but I had to make this change locally and then rerun. Do save me, and other people with crappy data, the suspense for the next time ! :)