Recognise more segfaults
I had a job die on a segfault, except it had an error code 139 and not 11.

I'm not sure if exit code 11 is special for treebest. I don't know if there are cases where the exit code is 11, the log doesn't say segfault, and it is a segfault, but this change preserves the existing behaviour in that scenario.