Updates to REST testing script
From e96, plants and vertebrates have a joint REST server and need different options to function.
Here, I've updated the script to include these options, along with a new defaults.

Updates were required for verification of XML leaves as structure differed.

Have also added an option to omit different sections of the testing. This can be used to switch off
sections that we know are not run (in the case of plants + EPO, for example), or it can be used to
switch off sections that have a known bug - allowing us to continue testing other sections easily.