Pipeline updated and documentation improved
The pipeline has been changed to create a new master database from
complete core databases (defined in the registry configuration file),
but includes the option to do the cloning of core regions if the
division includes the word "test". The documentation has been improved
and the parameters updated according to these changes.