bugfix: Do not try to copy the ncbi_taxa_name table again
"ncbi_taxa_name" is the onl table controlled by the "expand_tables" option.
The option controls whether to copy stuff that is not required by foreign
keys, e.g. taxon names.
Taxa are copied to satisfy the genome_db table. When this analysis runs,
all the genome_dbs are already in place, but the code still tries to copy
and expand the related data. Since there is no UNIQUE key in ncbi_taxa_name
the same rows as copied again and end up duplicated.