made dry_mode check the tickets to create/skip
Following @carlacummins suggestion, dry-run mode will now check which
tickets should be created and those already found on the JIRA server,
but they will not be created.

Additionally, _post_request() method checks if the password has been
already introduced, and if not, asks for it. This should not happen if
only the public methods are being used, but just in case.

The documentation has also been updated accordingly to these changes.