Ortheus.py is Python2 and can't be checked.
To check this file we would have to install another version fo pylint on
the side, and run it separately. mypy too can deal with Python2, but
with the `--py2` option.
Since we are not planning to add any other Python2 files, I felt it was
not worth making those changes, and decided to exclude Ortheus.py