Fix Sotrable against all perl versions.
This hotfix is temporary, a full fix will follow which has gone through
the PR process. It has been tested in similar form on live and in identical
form on a sandbox. Once it is pushed it will be retested on live.

Annoyingly this bug is both high-impact and only occurs in production which
leads to the odd process.