remove last become in our codebase
We control who can write in our shared spaces, e.g. $COMPARA_WAREHOUSE,
with the user/group permissions in linux, so there is no need any more
to become the virtual user: if you cannot write there, maybe it is
because you shouldn't (or maybe there is a bug, but better safe than
sorry).