Adaptation of species tree pipeline to use components
Apply filter for polyploid genomes with components such that genome dbs for the principals are excluded.
The largest change is a new method, filter_and_convert_genome_db_ids, in modules/Bio/EnsEMBL/Compara/Utils/DistanceMatrix.pm
Code for filtering before calling this method is added to runnables in modules/Bio/EnsEMBL/Compara/RunnableDB/SpeciesTree.
A change to the names used in the tree is made to include component information. Due to newick file format restrictions the use of display_name turns out to regularly cause errors and the parameter controlling this option has been removed.