Handle partial and complete liftover failures
Handle cases of complete liftover failure (where either the
halLiftover or axtChain output files are effectively empty)
and partial liftover failure (where only part of the source
region is lifted over, resulting in a mismatch between the
input source region and the lifted-over source region).

This moves the main script code to a main function, adds
a new high-level 'liftover_region' function, refactors
liftover post-processing, and reverts the StrandedRegion
class to a SimpleRegion.