bugfix: fix group_id from any source
`group_id` has the offset of the corresponding LASTZ_CHAIN MLSS id
instead of the LASTZ_NET MLSS id, and the code was breaking when trying
to check (and fix) the `group_id` of polyploid vs polyploid PWAs, where
there is one single LASTZ_NET but several LASTZ_CHAINs. Since the way we
handle `group_id` is internal, we have opted for not checking its range,
but just fixing it to be sure it is correct in the release database.