set 'default' collection for prot trees
For plants, metazoa and pan. In most cases this new collection is just a
copy of the division one, but leaves it ready for future changes as it
has happened with vertebrates and plants.

These changes eases the run of the protein trees pipelines, specially
when some parts needs to be rerun after it has completed (before, the
results were renamed from 'default' to the division name, having to
rename them to 'default' before running the analyses again).