bugfix: update metadata from master
instead of using the registry core dbs. This behaviour was working as
expected before freezing the list of species and we were in charge of
copying the species to vertannot. Since these two things have changed,
relying on the information in the registry can end on releasing species
that shouldn't, e.g. in e106 theobroma_cacao_criollo was being
"reactivated" because it was copied to vertannot when it shouldn't have.
Thus, the script now uses the current species in master to check
everything is correct.

Other changes:
- Updated POD
- Use current genomes to avoid reactivation of retired ones
- Removed division as it is not required if the information is based on
master db