Refactor main script code into functions
Put main code into functions 'main' and 'run_hal_liftover'.

Add unit tests for functions 'load_chr_sizes' and 'run_hal_liftover',
and add the flat files required for these tests.

Remove support for an input BED file and the dependency on pybedtools.
Rename the SimpleRegion reference sequence attribute from 'chrom' to
'chr' to be consistent with the rest of the code.

Update function 'make_src_region_file' to generate a more informative
error message in the event that the reference sequence of the given
region is not found in the input HAL file. Update its unit test to
include its new 'genome_name' parameter.