update config and requirements files
Most of these have been updated following the same ideas as in
ensembl-py. An attempt has been made to remove all upper version limits
in the libraries in `requirements.txt`. Furthermore, the minimum version
for `sqlalchemy` has been updated to 1.4.0 to match that of ensembl-py
(now a requirement, as well as two other Ensembl repositories).