add genomic_aln_mammoth analysis to ncRNA trees
As we did with MAFFT, several of the jobs (about 10%) required way more
than 32Gb to complete for default vertebrates e107. This new analysis
and flow should deal with these in a more automatic way.

NOTE: there will be still a few (~2%) that will require at least 128Gb
to complete.