Copy gene_member data before fetching GeneMember
In the ImportAltAlleGroupAsHomologies runnable, gene member data needs
to be copied to the gene_member table before a GeneMember object can be
instantiated. For a gene with a non-unique stable_id, data will be copied
for all gene members with that stable_id. If 'member_db' were to have many
non-unique stable_ids, this would become very wasteful. But for now, with
few or no non-unique stable_ids being present in practice, it's better to
have the data and not need it than need it and not have it.