Print original URL of MLSS if available
Printing a Cactus MLSS with a resolved base_dir-prefixed URL currently
requires that the $COMPARA_HAL_DIR environment variable is defined,
and that the resolved URL points to a file which exists in the current
machine. This seemed like quite an onerous requirement if all you want
to do is print the MLSS info to a string, so now the original URL is
printed instead.